--- .nyc_output/processinfo/1dcb6e55-a31f-4630-9826-62629453fe63.json
@@ -1,1 +1,0 @@
-{"parent":null,"pid":25432,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\node_modules\\mocha\\bin\\mocha.js","--timeout","20000"],"execArgv":[],"cwd":"C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models","time":1672769436313,"ppid":17716,"coverageFilename":"C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\.nyc_output\\1dcb6e55-a31f-4630-9826-62629453fe63.json","externalId":"","uuid":"1dcb6e55-a31f-4630-9826-62629453fe63","files":["C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\config.js","C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\scoring.js"]}
