+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"1dcb6e55-a31f-4630-9826-62629453fe63":{"parent":null,"children":[]}},"files":{"C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\config.js":["1dcb6e55-a31f-4630-9826-62629453fe63"],"C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\scoring.js":["1dcb6e55-a31f-4630-9826-62629453fe63"]},"externalIds":{}}
+{"processes":{"551678ff-de93-4118-9635-21701c4308ec":{"parent":null,"children":[]}},"files":{"C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\config.js":["551678ff-de93-4118-9635-21701c4308ec"],"C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\scoring.js":["551678ff-de93-4118-9635-21701c4308ec"],"C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\scoring_modules\\serializer.js":["551678ff-de93-4118-9635-21701c4308ec"],"C:\\Users\\cfernandezm\\Documents\\Developer\\Git\\models\\scoring_modules\\individuals.js":["551678ff-de93-4118-9635-21701c4308ec"]},"externalIds":{}}
