+++ package.json
@@ -25,7 +25,6 @@
     "body-parser": "^1.20.1",
     "csv-parser": "^3.0.0",
     "express": "^4.18.1",
-    "querystring": "^0.2.1",
     "simple-statistics": "^7.8.0",
     "winston": "^3.8.2",
     "winston-daily-rotate-file": "^4.7.1"
